+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This file keeps track of all changes to this project. This project follows [sema
 
 ## [UNRELEASED]
 
+### Added
+- Objective interfaces to reduce performance results for further analysis.
+- `BaseAnalyzer` to simply analyze a collection of configurations and retrieve all `PerformanceResult`s. 
+
 
 ## v0.2.1 - 2016-10-20